+++ Classes/Domain/Repository/QueryBuilder/TransferRepository.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace Brotkrueml\JobRouterData\Domain\Repository\QueryBuilder;
 
-use TYPO3\CMS\Core\Database\Query\QueryBuilder;
+use TYPO3\CMS\Core\Database\ConnectionPool;
 
 /**
  * @internal
@@ -20,11 +20,11 @@ class TransferRepository
 {
     private const TABLE_NAME = 'tx_jobrouterdata_domain_model_transfer';
 
-    private QueryBuilder $queryBuilder;
+    private ConnectionPool $connectionPool;
 
-    public function __construct(QueryBuilder $queryBuilder)
+    public function __construct(ConnectionPool $connectionPool)
     {
-        $this->queryBuilder = $queryBuilder;
+        $this->connectionPool = $connectionPool;
     }
 
     /**
@@ -32,7 +32,7 @@ class TransferRepository
      */
     public function countGroupByTransmitSuccess(): array
     {
-        $queryBuilder = $this->createQueryBuilder();
+        $queryBuilder = $this->connectionPool->getQueryBuilderForTable(self::TABLE_NAME);
 
         return $queryBuilder
             ->select('transmit_success')
@@ -45,7 +45,7 @@ class TransferRepository
 
     public function countTransmitFailed(): int
     {
-        $queryBuilder = $this->createQueryBuilder();
+        $queryBuilder = $this->connectionPool->getQueryBuilderForTable(self::TABLE_NAME);
 
         $whereExpressions = [
             $queryBuilder->expr()->eq(
@@ -74,7 +74,7 @@ class TransferRepository
 
     public function deleteOldSuccessfulTransfers(int $maximumTimestampForDeletion): int
     {
-        $queryBuilder = $this->createQueryBuilder();
+        $queryBuilder = $this->connectionPool->getQueryBuilderForTable(self::TABLE_NAME);
 
         return $queryBuilder
             ->delete(self::TABLE_NAME)
@@ -93,7 +93,7 @@ class TransferRepository
 
     public function findFirstCreationDate(): int
     {
-        $queryBuilder = $this->createQueryBuilder();
+        $queryBuilder = $this->connectionPool->getQueryBuilderForTable(self::TABLE_NAME);
 
         $quotedCrdate = $queryBuilder->quoteIdentifier('crdate');
 
@@ -102,10 +102,5 @@ class TransferRepository
             ->from(self::TABLE_NAME)
             ->execute()
             ->fetchColumn() ?: 0;
-    }
-
-    private function createQueryBuilder(): QueryBuilder
-    {
-        return clone $this->queryBuilder;
     }
 }